chore: rename __mock__ with mock
