+++ app/routes/favorites/index.server.test.ts
@@ -2,7 +2,7 @@ import { json } from '@remix-run/node';
 import getAnimalByIds from '~/models/Animal/getAnimalsByIds/index.server';
 import getContext from 'spec/utils/getContext';
 import getJsonFormData from 'spec/utils/getJsonFormData';
-import { ANIMALS } from 'spec/__mock__/constants/animal';
+import { ANIMALS } from 'spec/mock/constants/animal';
 import { action } from './index';
 
 jest.mock('~/models/animal/getAnimalsByIds/index.server');
