+++ spec/mock/components/AnimalCards/constants.ts
