+++ spec/mock/components/AnimalCards/index.tsx
