+++ spec/mock/constants/animal.ts
